meaningful error messages

+++ webpage/webpage_files/webpage_demo.py
@@ -13,8 +13,9 @@ To launch the webpage, enter in Ubuntu terminal:
 
 """
 
-from flask import Flask, render_template, request, send_from_directory
+from flask import Flask, flash, render_template, request, send_from_directory, redirect, url_for, Markup
 from werkzeug import secure_filename
+import requests
 import os
 import sys
 import json
@@ -24,10 +25,16 @@ homedir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 sys.path.append(homedir)
 from model_files.test_utils import load_model, single_prediction
 
+# Configuration parameters of the web application
 app = Flask(__name__)
 app.config['UPLOAD_FOLDER'] = 'uploads/'
-app.config['ALLOWED_EXTENSIONS'] = set(['png', 'jpg', 'jpeg'])
+app.config['ALLOWED_EXTENSIONS'] = set(['png', 'jpg', 'jpeg', 'PNG', 'JPG', 'JPEG'])
+if os.path.isfile('secret_key.txt'):
+    app.secret_key = open('secret_key.txt', 'r').read()
+else:
+    app.secret_key = 'devkey, should be in a file'
 
+# Loading image species  
 metadata = np.genfromtxt(os.path.join(homedir, 'model_files', 'data', 'synsets.txt'), dtype='str', delimiter='/n')
 
 # Load training info
@@ -49,7 +56,7 @@ label_list_to_html(os.path.join(homedir, 'model_files', 'data', 'synsets.txt'))
 
 
 @app.route('/')
-def my_form():
+def intmain():
     return render_template("main_page.html")
 
 
@@ -64,9 +71,29 @@ def static_from_root():
 
 
 @app.route('/url_upload', methods=['POST'])
-def my_form_post():
+def url_post():
     url = request.form['url']
     url = [i.replace(' ', '') for i in url.split(' ') if i != '']
+    
+    # Error catch: Empty query
+    if not url:
+        app.logger.error('Empty query')
+        error_message = Markup("""<center><b>Empty query</b></center>""")
+        flash(error_message)
+        return redirect(url_for('intmain'))
+        
+    # Error catch: Wrong formatted urls
+    for i in url:
+        url_type = requests.head(i).headers.get('content-type')
+        if url_type.split('/')[0] != 'image':
+            app.logger.error('Wrong url type (not image)')
+            error_message = Markup("""<center><b>Image formatting error</b></center><br>
+            Some urls were not in image format.
+            Check you didn't uploaded a preview of the image rather than the image itself.""")
+            flash(error_message)
+            return redirect(url_for('intmain'))
+    
+    # Predict
     pred_lab, pred_prob = single_prediction(test_func, im_list=url, aug_params={'mean_RGB': mean_RGB, 'filemode':'url'})
     return results_html_display(metadata[pred_lab], pred_prob)
 
@@ -82,6 +109,15 @@ def allowed_file(filename):
 @app.route('/local_upload', methods=['POST'])
 def local_post():
     uploaded_files = request.files.getlist("local_files")
+    
+    # Error catch: Empty query
+    if not uploaded_files[0].filename:
+        app.logger.error('Empty query')
+        error_message = Markup("""<center><b>Empty query</b></center>""")
+        flash(error_message)
+        return redirect(url_for('intmain'))   
+    
+    # Save images
     filenames = []
     for f in uploaded_files:
         if f and allowed_file(f.filename):
@@ -89,6 +125,16 @@ def local_post():
             file_path = os.path.join(homedir, 'webpage_files', 'templates', 'uploads', filename)
             f.save(file_path)
             filenames.append(file_path)
+            
+    # Error catch: Image format error
+    if not filenames:
+        app.logger.error('Local image format error')
+        error_message = Markup("""<center><b>Image formatting error</b></center>
+        Please use some common image format (jpg|jpeg|png).""")
+        flash(error_message)
+        return redirect(url_for('intmain'))            
+            
+    # Predict and clear images
     pred_lab, pred_prob = single_prediction(test_func, im_list=filenames, aug_params={'mean_RGB': mean_RGB, 'filemode':'local'})
     for f in filenames:
         os.remove(f)
